:truck: Renamed fadeIntext folder to fadeInText
 #281
